# split-bill-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AddMemberRequest:
    properties:
      nickname:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  models.CreateGroupRequest:
    properties:
      avatar_url:
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  models.GroupMemberResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      joined_at:
        type: string
      nickname:
        type: string
      role:
        $ref: '#/definitions/models.MemberRole'
      user_id:
        type: string
    type: object
  models.GroupResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      invite_code:
        type: string
      is_active:
        type: boolean
      members:
        items:
          $ref: '#/definitions/models.GroupMemberResponse'
        type: array
      name:
        type: string
    type: object
  models.JoinGroupRequest:
    properties:
      invite_code:
        type: string
    required:
    - invite_code
    type: object
  models.MemberRole:
    enum:
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
...
  models.UpdateGroupRequest:
    properties:
      avatar_url:
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@splitbill.app
    name: Split Bill Support
  description: API server for Split Bill application - manage groups, bills, transactions,
    OCR receipt scanning, and payments
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: 4.0.0
paths:
...
  /groups:
    get:
      description: Returns all groups the authenticated user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.GroupResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List user's groups
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Creates a new group with the authenticated user as creator and
        first member
      parameters:
      - description: Group creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - Groups
  /groups/{id}:
    delete:
      description: Deletes a group. Only the group creator can delete.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - Groups
    get:
      description: Returns detailed information about a specific group including member
        details
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GroupResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get group details
      tags:
      - Groups
    put:
      consumes:
      - application/json
      description: Updates group name, description, or currency. Only the group creator
        can update.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a group
      tags:
      - Groups
...
  /groups/{id}/members:
    post:
      consumes:
      - application/json
      description: Adds a user to a group by user ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Member data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Add member to group
      tags:
      - Groups
  /groups/{id}/members/{userId}:
    delete:
      description: Removes a user from a group. Only creator or the member themselves
        can remove.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove member from group
      tags:
      - Groups
...
  /groups/join:
    post:
      consumes:
      - application/json
      description: Joins a group using the group's invite code
      parameters:
      - description: Invite code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.JoinGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Join group via invite code
      tags:
      - Groups
...
securityDefinitions:
  BearerAuth:
    description: 'Firebase Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"